requirements.txt
asgiref==3.6.0
attrs==22.2.0
dj-database-url==1.2.0
Django==4.1.5
django-environ==0.9.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
psycopg2==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
whitenoise==6.3.0
